[JENKINS-55348] Re-write test to use MockSCM instead of mock GitHub's API
