--- src/test/java/jenkins/branch/naming/githubapimock/MockGithubJenkinsfileContent.java
@@ -1,17 +1,0 @@
-package jenkins.branch.naming.githubapimock;
-
-public class MockGithubJenkinsfileContent {
-    public MockGithubJenkinsfileContent() {}
-
-    public String getName() {
-        return "Jenkinsfile";
-    }
-
-    public String getPath() {
-        return "Jenkinsfile";
-    }
-
-    public String getType() {
-        return "file";
-    }
-}
