--- src/test/java/jenkins/branch/naming/githubapimock/MockGithubOrg.java
@@ -1,12 +1,0 @@
-package jenkins.branch.naming.githubapimock;
-
-public class MockGithubOrg {
-
-    public static final String ORG_LOGIN = "MyOrg";
-
-    public MockGithubOrg() {}
-
-    public String getLogin() {
-        return ORG_LOGIN;
-    }
-}
