--- src/test/java/jenkins/branch/naming/githubapimock/MockGithubRateLimit.java
@@ -1,9 +1,0 @@
-package jenkins.branch.naming.githubapimock;
-
-public class MockGithubRateLimit {
-    public MockGithubRateLimit() {}
-
-    public MockGithubResourcesRateLimit getResources() {
-        return new MockGithubResourcesRateLimit();
-    }
-}
