--- src/test/java/jenkins/branch/naming/githubapimock/MockGithubUser.java
@@ -1,12 +1,0 @@
-package jenkins.branch.naming.githubapimock;
-
-public class MockGithubUser {
-
-    public static final String USER_LOGIN = "JohnDoe";
-
-    public MockGithubUser() {}
-
-    public String getLogin() {
-        return USER_LOGIN;
-    }
-}
